fix(reservations): Reservation service can_client_create_reservation method always returning true and reservation recall button
This commit fixes the `ReservationService.can_client_create_reservation`
method always retruning true due the incorrect use of `not` on the
Result object instead of the real boolean result.

Besides that, it also fixes the button to keep on unfinished
reservations that was being showed even though the user had an active or
scheduled reservation.